+++ packages/@molt/command/src/types.ts
@@ -1,0 +1,1 @@
+export type SchemaBase = 'ZodBoolean' | 'ZodNumber' | 'ZodString'
