+++ packages/@molt/command/tests/environment.spec.ts
@@ -29,20 +29,20 @@ describe(`toggling`, () => {
     expect(() => Command.create({ '--foo': z.string() }).parseOrThrow([])).toThrow()
   })
   it(`can be enabled by settings`, () => {
-    environmentManager.set(`cli_foo`, `bar`)
+    environmentManager.set(`cli_param_foo`, `bar`)
     const args = Command.create({ '--foo': z.string() })
       .settings({ readArgumentsFromEnvironment: true })
       .parseOrThrow([])
     expect(args).toEqual({ foo: `bar` })
   })
   it(`can be enabled by environment`, () => {
-    environmentManager.set(`cli_environment_args`, `true`)
-    environmentManager.set(`cli_foo`, `bar`)
+    environmentManager.set(`CLI_SETTINGS_READ_ARGUMENTS_FROM_ENVIRONMENT`, `true`)
+    environmentManager.set(`cli_param_foo`, `bar`)
     const args = Command.create({ '--foo': z.string() }).parseOrThrow([])
     expect(args).toEqual({ foo: `bar` })
   })
   it(`environment supersedes settings`, () => {
-    environmentManager.set(`cli_environment_args`, `false`)
+    environmentManager.set(`CLI_SETTINGS_READ_ARGUMENTS_FROM_ENVIRONMENT`, `false`)
     environmentManager.set(`cli_foo`, `bar`)
     expect(() =>
       Command.create({ '--foo': z.string() })
@@ -52,55 +52,77 @@ describe(`toggling`, () => {
   })
 })
 
+describe(`default environment argument parameter name prefix`, () => {
+  beforeEach(() => environmentManager.set(`CLI_SETTINGS_READ_ARGUMENTS_FROM_ENVIRONMENT`, `true`))
+
+  it(`argument can be passed by CLI_PARAMETER prefix`, () => {
+    environmentManager.set(`cli_parameter_foo`, `bar`)
+    const args = Command.create({ '--foo': z.string() }).parseOrThrow([])
+    expect(args).toEqual({ foo: `bar` })
+  })
+  it(`argument can be passed by CLI_PARAM prefix`, () => {
+    environmentManager.set(`cli_param_foo`, `bar`)
+    const args = Command.create({ '--foo': z.string() }).parseOrThrow([])
+    expect(args).toEqual({ foo: `bar` })
+  })
+  it(`when both argument CLI_PARAM and CLI_PARAMETER are passed then an error is thrown`, () => {
+    environmentManager.set(`cli_param_foo`, `bar1`)
+    environmentManager.set(`cli_parameter_foo`, `bar2`)
+    expect(() => Command.create({ '--foo': z.string() }).parseOrThrow([])).toThrowErrorMatchingInlineSnapshot(
+      `"Multiple environment variables found for same parameter \\"foo\\": [object Object], [object Object]"`
+    )
+  })
+})
+
 describe(`when enabled and a flag arg is not passed then the env is considered`, () => {
-  beforeEach(() => environmentManager.set(`CLI_ENV_ARGS`, `true`))
+  beforeEach(() => environmentManager.set(`CLI_SETTINGS_READ_ARGUMENTS_FROM_ENVIRONMENT`, `true`))
 
   describe(`boolean`, () => {
     it(`true`, () => {
-      environmentManager.set(`cli_VERBOSE`, `true`)
+      environmentManager.set(Command.environmentArgumentName(`VERBOSE`), `true`)
       const args = Command.create({ '--verbose': z.boolean() }).parseOrThrow([])
       expect(args).toEqual({ verbose: true })
     })
     it(`true (with param default false)`, () => {
-      environmentManager.set(`cli_VERBOSE`, `true`)
+      environmentManager.set(Command.environmentArgumentName(`VERBOSE`), `true`)
       const args = Command.create({ '--verbose': z.boolean().default(false) }).parseOrThrow([])
       expect(args).toEqual({ verbose: true })
     })
     it(`false`, () => {
-      environmentManager.set(`cli_verbose`, `false`)
+      environmentManager.set(Command.environmentArgumentName(`verbose`), `false`)
       const args = Command.create({ '--verbose': z.boolean() }).parseOrThrow([])
       expect(args).toEqual({ verbose: false })
     })
   })
 
   it(`string`, () => {
-    environmentManager.set(`cli_foo`, `bar`)
+    environmentManager.set(Command.environmentArgumentName(`foo`), `bar`)
     const args = Command.create({ '--foo': z.string() }).parseOrThrow([])
     expect(args).toEqual({ foo: `bar` })
   })
   it(`number`, () => {
-    environmentManager.set(`cli_foo`, `4.3`)
+    environmentManager.set(Command.environmentArgumentName(`foo`), `4.3`)
     const args = Command.create({ '--foo': z.number() }).parseOrThrow([])
     expect(args).toEqual({ foo: 4.3 })
   })
   it(`env arg is validated`, () => {
-    environmentManager.set(`cLi_fOo`, `d`)
+    environmentManager.set(Command.environmentArgumentName(`foo`), `d`)
     expect(() => Command.create({ '--foo': z.enum([`a`, `b`, `c`]) }).parseOrThrow([]))
       .toThrowErrorMatchingInlineSnapshot(`
-        "Invalid argument (via environment variable \\"CLI_FOO\\") for parameter: \\"foo\\". The error was:
+        "Invalid argument (via environment variable \\"CLI_PARAMETER_FOO\\") for parameter: \\"foo\\". The error was:
         Invalid enum value. Expected 'a' | 'b' | 'c', received 'd'"
       `)
   })
   it(`case of env name does not matter`, () => {
-    environmentManager.set(`cLi_fOo`, `bar`)
+    environmentManager.set(`cLi_PARAM_fOo`, `bar`)
     const args = Command.create({ '--foo': z.string() }).parseOrThrow([])
     expect(args).toEqual({ foo: `bar` })
   })
 })
 
-it(`if environment args enabled, parameter has default, flag arg not given, but env arg given, then env arg wins`, () => {
-  environmentManager.set(`CLI_ENV_ARGS`, `true`)
-  process.env[`cli_verbose`] = `true`
-  const args = Command.create({ '--verbose': z.boolean().default(false) }).parseOrThrow([])
-  expect(args).toEqual({ verbose: true })
-})
+// it(`if environment args enabled, parameter has default, flag arg not given, but env arg given, then env arg wins`, () => {
+//   environmentManager.set(`CLI_SETTINGS_READ_ARGUMENTS_FROM_ENVIRONMENT`, `true`)
+//   process.env[`cli_verbose`] = `true`
+//   const args = Command.create({ '--verbose': z.boolean().default(false) }).parseOrThrow([])
+//   expect(args).toEqual({ verbose: true })
+// })
